# online_exam_system_vue/OpenAPI-DOC.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9999",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/exams": {
      "post": {
        "tags": [
          "考试管理"
        ],
        "summary": "创建考试",
        "description": "创建一个新的考试",
        "operationId": "createExam",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Exams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/exams/random": {
      "post": {
        "tags": [
          "考试管理"
        ],
        "summary": "随机组卷",
        "description": "根据题型分布自动生成考试试卷",
        "operationId": "createRandomExam",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RandomExamDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/exams/{page}/{size}": {
      "get": {
        "tags": [
          "考试管理"
        ],
        "summary": "分页获取考试列表",
        "description": "分页获取考试列表",
        "operationId": "getExams",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "页码",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "path",
            "description": "每页大小",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/exams/{examId}/details": {
      "get": {
        "tags": [
          "考试管理"
        ],
        "summary": "获取考试详情",
        "description": "根据考试ID获取考试的详细信息，包括题目列表",
        "operationId": "getExamDetails",
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "description": "考试ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "授权令牌",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/exams/student/{studentId}/{page}/{size}": {
      "get": {
        "tags": [
          "考试管理"
        ],
        "summary": "获取学生考试列表",
        "description": "根据学生ID分页获取该学生的考试列表",
        "operationId": "getExamsByStudentId",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "description": "学生ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "path",
            "description": "页码",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "path",
            "description": "每页大小",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/exams/search": {
      "get": {
        "tags": [
          "考试管理"
        ],
        "summary": "搜索考试",
        "description": "根据关键字搜索考试，支持分页",
        "operationId": "searchExams",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "搜索关键字",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页大小",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/exam-submissions/exam/{examId}": {
      "get": {
        "tags": [
          "考试提交管理"
        ],
        "summary": "获取考试提交记录",
        "description": "根据考试ID获取该考试的所有提交记录",
        "operationId": "getExamSubmissionsByExam",
        "parameters": [
          {
            "name": "examId",
            "in": "path",
            "description": "考试ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExamSubmissionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/exams/{id}": {
      "delete": {
        "tags": [
          "考试管理"
        ],
        "summary": "删除考试",
        "description": "根据考试ID删除指定考试",
        "operationId": "deleteExam",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "考试ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Exams": {
        "required": [
          "courseId",
          "creatorId",
          "endTime",
          "needsGrading",
          "startTime",
          "title",
          "totalScore"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "creatorId": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "totalScore": {
            "type": "integer",
            "format": "int32"
          },
          "questions": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "courseId": {
            "type": "integer",
            "format": "int64"
          },
          "needsGrading": {
            "type": "boolean"
          }
        }
      },
      "QuestionDistribution": {
        "required": [
          "count",
          "scorePerQuestion",
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "difficulty": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "scorePerQuestion": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "RandomExamDto": {
        "required": [
          "courseId",
          "creatorId",
          "endTime",
          "questionDistribution",
          "startTime",
          "title",
          "totalScore"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "creatorId": {
            "type": "integer",
            "format": "int64"
          },
          "totalScore": {
            "type": "integer",
            "format": "int32"
          },
          "courseId": {
            "type": "integer",
            "format": "int64"
          },
          "questionDistribution": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionDistribution"
            }
          }
        }
      },
...
      "ExamSubmissionResponse": {
        "required": [
          "examId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "examId": {
            "type": "integer",
            "format": "int64"
          },
          "userName": {
            "type": "string"
          },
          "examTitle": {
            "type": "string"
          },
          "examTotalScore": {
            "type": "integer",
            "format": "int32"
          },
          "submissionScore": {
            "type": "integer",
            "format": "int32"
          },
          "answers": {
            "type": "object"
          }
        }
      },
...
    }
  },
...
}